Fix CI: examples deployment. (#38)
The step that builds the examples fails because the UI static files are not included in the source code anymore (see #37).
In order to fix it, we need to compile the UI before running the examples. To avoid repeating the steps, a new reusable workflow named `compile-ui.yml` is created that can be called from the other workflows.